Добавление типа переплета
