+++ react-web-client/src/components/admin-panel/cover-type/add-cover-type/AddCoverType.css
