+++ react-web-client/src/render.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import { setPainterStyle, initializeData, deletePainterStyle, updatePainterStyle, setPainter, deletePainter, updatePainter, getPainterByCount, addPublisher, getPublishersByCount, deletePublisher, updatePublisher, getCoverTypes } from './redux/state';
+import { setPainterStyle, initializeData, deletePainterStyle, updatePainterStyle, setPainter, deletePainter, updatePainter, getPainterByCount, addPublisher, getPublishersByCount, deletePublisher, updatePublisher, getCoverTypes, addCoverType } from './redux/state';
 
 export let rerenderEntireTree = (state) => {
   let painterMethods = {
@@ -20,13 +20,14 @@ export let rerenderEntireTree = (state) => {
     updatePublisher
   }
 
-  let coverTypesMethods = {
-    getCoverTypes
+  let coverTypeMethods = {
+    getCoverTypes,
+    addCoverType
   }
 
   ReactDOM.render(
     <React.StrictMode>
-      <App state={state} setPainterStyle={setPainterStyle} deletePainterStyle={deletePainterStyle} updatePainterStyle={updatePainterStyle} painterMethods={painterMethods} publisherMethods={publisherMethods} coverTypesMethods={coverTypesMethods}/>
+      <App state={state} setPainterStyle={setPainterStyle} deletePainterStyle={deletePainterStyle} updatePainterStyle={updatePainterStyle} painterMethods={painterMethods} publisherMethods={publisherMethods} coverTypeMethods={coverTypeMethods}/>
     </React.StrictMode>,
     document.getElementById('root')
   );
